rename format to templates
